move RppgResult and Color into separate file
