+++ src/yarppg/processors/processor.py
@@ -1,56 +1,16 @@
 """Base processor for rPPG signal computation."""
 
-from dataclasses import dataclass
-
 import cv2
 import numpy as np
 
 from ..roi.region_of_interest import RegionOfInterest
-
-
-@dataclass
-class Color:
-    """Defines a color in RGB(A) format."""
-
-    r: float
-    g: float
-    b: float
-    a: float = 1.0
-
-    @classmethod
-    def null(cls):
-        """Create empty color with NaN values."""
-        return cls(np.nan, np.nan, np.nan)
-
-    def __array__(self):
-        return np.array([self.r, self.g, self.b, self.a])
-
-    @classmethod
-    def from_array(cls, arr: np.ndarray):
-        """Convert numpy array to `Color` object."""
-        if len(arr) in {3, 4} and arr.ndim == 1:
-            return cls(*arr)
-        raise ValueError(f"Cannot interpret {arr=!r}")
+from ..rppg_result import Color, RppgResult
 
 
 def masked_average(frame: np.ndarray, mask: np.ndarray) -> Color:
     """Calculate average color of the masked region."""
     r, g, b, a = cv2.mean(frame, mask)
     return Color(r, g, b, a)
-
-
-@dataclass
-class RppgResult:
-    """Container for rPPG computation results."""
-
-    value: float
-    roi: RegionOfInterest
-    roi_mean: Color
-    bg_mean: Color
-    hr: float = np.nan
-
-    def __array__(self):
-        return np.asarray([self.value])
 
 
 class Processor:
