+++ src/yarppg/roi/__init__.py
@@ -1,6 +1,6 @@
 """Utilities for ROI (region of interest) detection and manipulation."""
 
-from detector import ROIDetector
-from facemesh_segmenter import FaceMeshDetector
-from region_of_interest import RegionOfInterest, pixelate_mask
-from selfie_segmenter import SelfieDetector
+from .detector import ROIDetector
+from .facemesh_segmenter import FaceMeshDetector
+from .region_of_interest import RegionOfInterest, pixelate_mask
+from .selfie_segmenter import SelfieDetector
